requirements.txt
aniso8601==9.0.1
cachelib==0.1.1
cffi==1.14.6
click==7.1.2
cryptography==35.0.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-RESTful==0.3.8
Flask-Session==0.3.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
nltk==3.5
protobuf==3.15.6
pycparser==2.20
pydub==0.25.1
PyJWT==2.1.0
PyMySQL==1.0.2
pytz==2021.1
regex==2021.3.17
six==1.15.0
SpeechRecognition==3.8.1
tqdm==4.59.0
Werkzeug==1.0.1
